Crossbowtower

+++ src/game/scenes/GameScene.java
@@ -71,7 +71,7 @@ public class GameScene extends Scene
         //addObject(new TestMouseCollider()); // debug
     }
     
-    public void damage(int amount)
+    public void damage(double amount)
     {
         health -= amount;
         if(health <= 0)
@@ -100,7 +100,7 @@ public class GameScene extends Scene
             layer.graphics().setColor(Color.YELLOW);
             layer.drawStringCentered((int)money+"$", MENU_AREA_START.add(MENU_AREA_SIZE.x / 2 + 20, 20));
             layer.graphics().setColor(Color.RED);
-            layer.drawStringCentered(health + "♥", MENU_AREA_START.add(40, 20));
+            layer.drawStringCentered((int)health + "♥", MENU_AREA_START.add(40, 20));
             
         }
